pyproject.toml
[project]
name = "SortingMoney"
version = "1.0"
dependencies = [
    "pydantic>=2.10.2",
    "pyyaml>=6.0.2",
    "rich>=13.9.4",
    "textual>=0.89.1",
    "typer>=0.15.1",
]
requires-python = ">=3.12"

[project.scripts]
moneymanager = "moneymanager.__main__:app"

[tool.uv]
dev-dependencies = [
    # "uv",
    "pyright",
    "debugpy",
    "ruff",
    "tox",
    "tox-uv",
]

[tool.setuptools]
packages = ["moneymanager"]

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py312

[testenv]
runner = uv-venv-lock-runner
description = run all checks on the code
extras = 
    dev
commands =
    ruff format --check src
    ruff check src
    pyright src/
"""

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py312"
src = ["src"]
exclude = ["bin"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle Error
    "F",     # pyflakes
    "UP",    # pyupgrade
    "SIM",   # flake8-simplify
    "I",     # imports
    "S",     # bandit (security)
    "N",     # pep8-naming
    "ASYNC", # flake8-async
    "C4",    # flake8-compehensions
    "FA",    # flake8-future-annotations
    "ISC",   # flake8-implicit-str-concat
    "ICN",   # flake8-import-conventions
    "G",     # flake8-logging-format
    "PIE",   # flake8-pie
    "PYI",   # flake8-pyi
    "RSE",   # flake8-raise
    "SLOT",  # flake8-slots
    "INT",   # flake8-gettext
    "TRY",   # tryceratops
    "FLY",   # flynt
    "PERF",  # Perflint
    "FURB",  # refurb
    "LOG",   # flake8-logging
    "RUF",   # Ruff-specific-rules
    # "ERA",   # locate commented codes
    # "FIX",   # locate TODOs and FIXME
    # "PTH",   # flake8-use-pathlib (maybe todo ?)
    # "TID",   # flake8-tidy-imports (maybe todo ?)
    # "SLF",   # flake8-self  (managed by pyright)
    # "RET",   # flake8-return
    # "Q",     # flake8-quotes
    # "T20",   # flake8-print
    # "DTZ",   # flake8-datetimez (TODO)
    # "B",     # flake8-bugbear
]
ignore = [
    "E501", # line too long (we relate on the formater)
    "N818", # Error suffix for exceptions names
    # "PIE796", # Enum contains duplicate value
    "TRY003", # Avoid specifying long messages outsides the exception class
    "ISC001", # To avoid conflicts with the formatter
    "TRY002",
]
dummy-variable-rgx = '^\*{0,2}(_$|__$|unused_|dummy_)'

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.lint.pyflakes]
extend-generics = ["discord.app_commands.CommandTree"]
